Added some details to installation instructions
Hi! SendCode adds amazing functionality to the sublime text editor and I've been really happy using it. I was a little inexperienced with using Package Control for installation, and the current instructions were a little brief so I had to do a fair bit of googling to complete the install. I've added some detail to help people less familiar with the process get SendCode up and running.